requirements.txt
asgiref==3.8.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.8
psycopg==3.2.1
PyJWT==2.9.0
pysrt==1.1.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
youtube-dl==2021.12.17
